fixed error demo

+++ TiposPrimitivos_Wrapper/inteiro_positivo/InteiroPositivo.java
@@ -1,8 +1,8 @@
 public class InteiroPositivo {
     
-    private float valor;
+    private int valor;
 
-    public InteiroPositivo(float valor) {
+    public InteiroPositivo(int valor) {
         setValor(valor);
     }
